requirements.txt
# Core dependencies
pydantic>=2.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
openai>=1.0.0

# Configuration and utilities
PyYAML>=6.0
python-dotenv>=1.0.0

# Async support
asyncio
aiofiles

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: MCP support (when available)
# mcp-client  # Uncomment when MCP Python SDK is available

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0 